backend/requirements.txt
sentence-transformers
pymongo
fastapi
uvicorn
pymongo
faiss-cpu
openai
python-dotenv
anthropic